Move AuthUserModel to shared lib to prevent circular dependency
