+++ libs/gamify/auth/src/index.ts
@@ -1,4 +1,3 @@
 export * from './lib/auth.module';
 export * from './lib/guards';
 export * from './lib/decorators';
-export * from './lib/models';
