--- libs/gamify/auth/src/lib/models/index.ts
@@ -1,1 +1,0 @@
-export * from './auth-user.model';
