+++ libs/gamify/shared/src/lib/models/index.ts
@@ -3,3 +3,4 @@ export * from './badge.model';
 export * from './user.model';
 export * from './application-user.model';
 export * from './user-badge.model';
+export * from './auth-user.model';
